+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Margins added to stylesheet system. Margins automatically handled for all widgets.
 - Added `widget::input::State::set_value`
 - Removed scrollbars from stylesheet in favor of the new `Dummy` widget.
+- Added `Progress` widget.
 
  ### v0.3.0
 - Added `len()` to `Widget`.